+++ src/temporarySettingsManager.ts
@@ -63,9 +63,12 @@ export class TemporarySettingsManager implements Disposable {
 
   public register(): void {
     this.disposables.push(
-      commands.registerCommand(resetTemporarySettingsCommandID, async () => {
-        await this.resetSettings();
-      })
+      commands.registerCommand(
+        "vscode-as-git-mergetool.resetTemporarySettings",
+        async () => {
+          await this.resetSettings();
+        }
+      )
     );
   }
 
@@ -116,5 +119,3 @@ type StorageState = { [k: string]: unknown };
 
 const origActualKey = `${extensionID}.temporarySettings.origActual`;
 const origTargetKey = `${extensionID}.temporarySettings.origTarget`;
-const resetTemporarySettingsCommandID =
-  "vscode-as-git-mergetool.resetTemporarySettings";